[refactor] Move test folder from `src` to root folder
