+++ prosemirror-jdita/.mocharc.json
@@ -1,5 +1,5 @@
 {
   "extension": ["ts"],
   "require": "ts-node/register",
-  "spec": "src/**/*.spec.ts"
+  "spec": "tests/**/*.spec.ts"
 }
