+++ prosemirror-jdita/tests/document.spec.ts
@@ -2,7 +2,7 @@ import ChaiPromised from 'chai-as-promised';
 import { use } from 'chai';
 import { expect, assert } from 'chai';
 import { xditaToJson } from 'jdita';
-import { document } from './document';
+import { document } from '../src/document';
 
 use(ChaiPromised);
