+++ prosemirror-jdita/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "exclude": [
     "**/*.spec.ts",
-    "src/tests.ts"
+    "tests/*.ts"
   ],
   "compilerOptions": {
     /* Basic Options */
